+++ packages/pkg-core/src/commands/index.ts
@@ -1,1 +1,1 @@
-export {BaseCommand, handleCommandError} from './base-command'
+export {BaseCommand} from './base-command'